move create_accent_phrases to synthesis_engine
