+++ voicevox_engine/synthesis_engine.py
@@ -8,6 +8,8 @@ import resampy
 
 from voicevox_engine.acoustic_feature_extractor import OjtPhoneme, SamplingData
 from voicevox_engine.model import AccentPhrase, AudioQuery, Mora
+from voicevox_engine.full_context_label import extract_full_context_label
+from voicevox_engine.mora_list import openjtalk_mora2text
 
 unvoiced_mora_phoneme_list = ["A", "I", "U", "E", "O", "cl", "pau"]
 mora_phoneme_list = ["a", "i", "u", "e", "o", "N"] + unvoiced_mora_phoneme_list
@@ -123,6 +125,16 @@ def pre_process(
     phoneme_data_list = to_phoneme_data_list(phoneme_str_list)
 
     return flatten_moras, phoneme_data_list
+
+
+def mora_to_text(mora: str):
+    if mora[-1:] in ["A", "I", "U", "E", "O"]:
+        # 無声化母音を小文字に
+        mora = mora[:-1] + mora[-1].lower()
+    if mora in openjtalk_mora2text:
+        return openjtalk_mora2text[mora]
+    else:
+        return mora
 
 
 class SynthesisEngine:
@@ -364,6 +376,59 @@ class SynthesisEngine:
                 accent_phrases=accent_phrases,
                 speaker_id=speaker_id,
             ),
+            speaker_id=speaker_id,
+        )
+
+    def create_accent_phrases(self, text: str, speaker_id: int) -> List[AccentPhrase]:
+        if len(text.strip()) == 0:
+            return []
+
+        utterance = extract_full_context_label(text)
+        if len(utterance.breath_groups) == 0:
+            return []
+
+        return self.replace_mora_data(
+            accent_phrases=[
+                AccentPhrase(
+                    moras=[
+                        Mora(
+                            text=mora_to_text(
+                                "".join([p.phoneme for p in mora.phonemes])
+                            ),
+                            consonant=(
+                                mora.consonant.phoneme
+                                if mora.consonant is not None
+                                else None
+                            ),
+                            consonant_length=0 if mora.consonant is not None else None,
+                            vowel=mora.vowel.phoneme,
+                            vowel_length=0,
+                            pitch=0,
+                        )
+                        for mora in accent_phrase.moras
+                    ],
+                    accent=accent_phrase.accent,
+                    pause_mora=(
+                        Mora(
+                            text="、",
+                            consonant=None,
+                            consonant_length=None,
+                            vowel="pau",
+                            vowel_length=0,
+                            pitch=0,
+                        )
+                        if (
+                            i_accent_phrase == len(breath_group.accent_phrases) - 1
+                            and i_breath_group != len(utterance.breath_groups) - 1
+                        )
+                        else None
+                    ),
+                )
+                for i_breath_group, breath_group in enumerate(utterance.breath_groups)
+                for i_accent_phrase, accent_phrase in enumerate(
+                    breath_group.accent_phrases
+                )
+            ],
             speaker_id=speaker_id,
         )
